refactor: View -> AbstractView 변경, 접근 제어자 default로 설정
