+++ src/main/java/webserver/http/servlet/AbstractView.java
@@ -1,0 +1,8 @@
+package webserver.http.servlet;
+
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
+
+public abstract class AbstractView {
+    abstract void render(HttpRequest request, HttpResponse response);
+}
