+++ src/main/java/webserver/http/servlet/FileView.java
@@ -1,4 +1,4 @@
-package webserver.http.servlet.view;
+package webserver.http.servlet;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,7 +7,7 @@ import webserver.http.request.HttpRequest;
 import webserver.http.request.RequestURI;
 import webserver.http.response.HttpResponse;
 
-public class FileView implements View {
+public class FileView extends AbstractView {
     private static final Logger logger = LoggerFactory.getLogger(FileView.class);
 
     private final RequestURI uri;
@@ -17,7 +17,7 @@ public class FileView implements View {
     }
 
     @Override
-    public void render(HttpRequest request, HttpResponse response) {
+    void render(HttpRequest request, HttpResponse response) {
         try {
             response.ok(uri.getClassPath(), uri.getContentType());
         } catch (IllegalArgumentException e) {
