+++ src/main/java/webserver/http/servlet/HttpServlet.java
@@ -5,7 +5,6 @@ import webserver.http.request.RequestMapping;
 import webserver.http.response.HttpResponse;
 import webserver.http.servlet.controller.Controller;
 import webserver.http.servlet.controller.ControllerMapping;
-import webserver.http.servlet.view.View;
 
 public class HttpServlet {
     private static final HttpServlet HTTP_SERVLET = new HttpServlet();
@@ -20,7 +19,7 @@ public class HttpServlet {
     public void doDispatch(HttpRequest request, HttpResponse response) {
         RequestMapping requestMapping = request.toRequestMapping();
         Controller controller = ControllerMapping.get(requestMapping);
-        View view = controller.doService(request, response);
+        AbstractView view = controller.doService(request, response);
         view.render(request, response);
     }
 }
