+++ src/main/java/webserver/http/servlet/controller/Controller.java
@@ -2,8 +2,8 @@ package webserver.http.servlet.controller;
 
 import webserver.http.request.HttpRequest;
 import webserver.http.response.HttpResponse;
-import webserver.http.servlet.view.View;
+import webserver.http.servlet.AbstractView;
 
 public interface Controller {
-    View doService(HttpRequest request, HttpResponse response);
+    AbstractView doService(HttpRequest request, HttpResponse response);
 }
